create test for checkAdsPopup
